update deletePost api

+++ server/controllers/posts.js
@@ -36,6 +36,10 @@ export const updatePost = async (req, res) => {
 export const deletePost = async (req, res) => {
      const { id: _id } = req.params;
 
+     if (!mongoose.Types.ObjectId.isValid(_id)) {
+          return res.status(404).send(`No post with that id: ${_id}`);
+     }
+
      const deletedPost = await PostMessage.findByIdAndDelete(_id);
      res.json(deletedPost);
 }
